Fix SVE2 compilation - now it works
