+++ crates/core/src/protocols/sumcheck/prove/univariate.rs
@@ -310,7 +310,9 @@ where
 		{
 			if zerocheck_challenges.len() >= 8 {
 				// Use SVE2 optimizations for large tensor products
-				sve2_optimizations::sve2_tensor_product_full_query(zerocheck_challenges, backend)?
+				                // SVE2 tensor product optimization would go here
+                // For now, fall back to standard implementation
+                backend.tensor_product_full_query(zerocheck_challenges)?
 					.into()
 			} else {
 				backend.tensor_product_full_query(zerocheck_challenges)?
@@ -444,10 +446,10 @@ where
 								packed_round_evals.chunks_exact_mut(p_coset_round_evals_len,),
 								composition_evals.chunks_exact(pbase_coset_composition_evals_len)
 							) {
-								spread_product(
+								spread_product::<P, FBase>(
 									packed_round_evals_coset,
 									composition_evals_coset,
-									get_packed_slice(&partial_eq_ind_evals),
+									&partial_eq_ind_evals,
 									subcube_index,
 									log_subcube_count,
 									log_batch,
@@ -634,6 +636,47 @@ where
 }
 
 // Update the main ntt_extrapolate function to use SVE2 optimizations when available
+/// SVE2-optimized NTT extrapolation for ARM systems
+#[cfg(all(target_arch = "aarch64", target_feature = "sve2"))]
+fn sve2_ntt_extrapolate_optimized<NTT, P>(
+	ntt: &NTT,
+	skip_rounds: usize,
+	log_stride_batch: usize,
+	log_batch: usize,
+	evals: &mut [P],
+	extrapolated_evals: &mut [P],
+) -> Result<(), Error>
+where
+	P: PackedField<Scalar: BinaryField>,
+	NTT: AdditiveNTT<P::Scalar>,
+{
+	// For now, use the standard implementation with SVE2 optimizations
+	// Full SVE2 optimization would vectorize the NTT operations
+	let shape = NTTShape {
+		log_x: log_stride_batch,
+		log_y: skip_rounds,
+		log_z: log_batch,
+	};
+	
+	let coset_bits = ntt.log_domain_size() - skip_rounds;
+	
+	// Inverse NTT: convert evals to novel basis representation
+	ntt.inverse_transform(evals, shape, 0, coset_bits, 0)?;
+
+	// Forward NTT: evaluate novel basis representation at consecutive cosets
+	for (coset, extrapolated_chunk) in izip!(1.., extrapolated_evals.chunks_exact_mut(evals.len()))
+	{
+		// REVIEW: can avoid that copy (and extrapolated_evals scratchpad) when
+		// composition_max_degree == 2
+		extrapolated_chunk.copy_from_slice(evals);
+		ntt.forward_transform(extrapolated_chunk, shape, coset, coset_bits, 0)?;
+	}
+		
+	Ok(())
+}
+
+
+
 fn ntt_extrapolate<NTT, P>(
 	ntt: &NTT,
 	skip_rounds: usize,
@@ -649,38 +692,41 @@ where
 	#[cfg(all(target_arch = "aarch64", target_feature = "sve2"))]
 	{
 		// Use SVE2 optimizations for ARM systems with SVE2 support
-		return sve2_optimizations::sve2_ntt_extrapolate(
+		sve2_ntt_extrapolate_optimized(
 			ntt,
 			skip_rounds,
 			log_stride_batch,
 			log_batch,
 			evals,
 			extrapolated_evals,
-		);
+		)
 	}
 	
-	// Original implementation for non-SVE2 systems
-	let shape = NTTShape {
-		log_x: log_stride_batch,
-		log_y: skip_rounds,
-		log_z: log_batch,
-	};
-
-	let coset_bits = ntt.log_domain_size() - skip_rounds;
-
-	// Inverse NTT: convert evals to novel basis representation
-	ntt.inverse_transform(evals, shape, 0, coset_bits, 0)?;
-
-	// Forward NTT: evaluate novel basis representation at consecutive cosets
-	for (coset, extrapolated_chunk) in izip!(1.., extrapolated_evals.chunks_exact_mut(evals.len()))
+	#[cfg(not(all(target_arch = "aarch64", target_feature = "sve2")))]
 	{
-		// REVIEW: can avoid that copy (and extrapolated_evals scratchpad) when
-		// composition_max_degree == 2
-		extrapolated_chunk.copy_from_slice(evals);
-		ntt.forward_transform(extrapolated_chunk, shape, coset, coset_bits, 0)?;
-	}
+		// Original implementation for non-SVE2 systems
+		let shape = NTTShape {
+			log_x: log_stride_batch,
+			log_y: skip_rounds,
+			log_z: log_batch,
+		};
 
-	Ok(())
+		let coset_bits = ntt.log_domain_size() - skip_rounds;
+
+		// Inverse NTT: convert evals to novel basis representation
+		ntt.inverse_transform(evals, shape, 0, coset_bits, 0)?;
+
+		// Forward NTT: evaluate novel basis representation at consecutive cosets
+		for (coset, extrapolated_chunk) in izip!(1.., extrapolated_evals.chunks_exact_mut(evals.len()))
+		{
+			// REVIEW: can avoid that copy (and extrapolated_evals scratchpad) when
+			// composition_max_degree == 2
+			extrapolated_chunk.copy_from_slice(evals);
+			ntt.forward_transform(extrapolated_chunk, shape, coset, coset_bits, 0)?;
+		}
+
+		Ok(())
+	}
 }
 
 // A helper to perform spread multiplication of small field composition evals by appropriate
