+++ crates/field/src/arch/aarch64/simd_arithmetic.rs
@@ -461,37 +461,10 @@ fn lagrange_interpolate_single(points: &[M128], values: &[M128], eval_point: M12
 #[cfg(all(target_arch = "aarch64", target_feature = "sve"))]
 #[inline]
 pub fn lookup_sve_optimized(table_neon: &[uint8x16x4_t; 4], x: M128) -> M128 {
-	unsafe {
-		use std::arch::asm;
-		
-		// Convert NEON table to SVE-compatible format
-		let mut result = x.into();
-		
-		// SVE table lookup with predicated operations
-		// This uses SVE's scalable vector capabilities for better performance
-		asm!(
-			// Load input vector into SVE register
-			"ptrue p0.b",                           // Set all predicate bits
-			"dup z0.b, {input:w}",                  // Duplicate input across SVE vector
-			
-			// Perform table lookups with SVE
-			"tbl z1.b, {{v{t0}.16b, v{t1}.16b, v{t2}.16b, v{t3}.16b}}, z0.b",
-			
-			// Extract result back to NEON
-			"mov {result:x}, z1.d[0]",
-			
-			input = in(reg) result as u8,
-			t0 = in(vreg) table_neon[0],
-			t1 = in(vreg) table_neon[1], 
-			t2 = in(vreg) table_neon[2],
-			t3 = in(vreg) table_neon[3],
-			result = out(reg) result,
-			options(pure, nomem, nostack)
-		);
-		
-		// Fall back to NEON for compatibility if SVE assembly fails
-		lookup_16x8b_neon_precomputed(table_neon, x)
-	}
+	// For now, fall back to optimized NEON implementation
+	// Full SVE table lookup requires restructuring the lookup tables
+	// which is beyond the scope of current optimizations
+	lookup_16x8b_neon_precomputed(table_neon, x)
 }
 
 // Ultra-fast NEON lookup using pre-computed table format
@@ -518,48 +491,17 @@ pub fn lookup_16x8b_neon_precomputed(table_neon: &[uint8x16x4_t; 4], x: M128) ->
 #[inline]
 pub fn packed_tower_sve_multiply(a: M128, b: M128) -> M128 {
 	unsafe {
-		use std::arch::asm;
+		// Fast path for zero operands using NEON
+		let a_vec: uint8x16_t = a.into();
+		let b_vec: uint8x16_t = b.into();
+		let a_is_zero = vceqzq_u8(a_vec);
+		let b_is_zero = vceqzq_u8(b_vec);
+		let either_zero = vorrq_u8(a_is_zero, b_is_zero);
 		
-		// Fast path for zero operands using SVE predicate operations
-		let a_vec = a.into();
-		let b_vec = b.into();
-		
-		let mut result: u128;
-		
-		// SVE-optimized logarithmic multiplication
-		asm!(
-			// Load vectors into SVE registers
-			"ptrue p0.b",                           // All lanes active
-			"mov z0.d, {a_vec:x}",                  // Load a into SVE
-			"mov z1.d, {b_vec:x}",                  // Load b into SVE
-			
-			// Check for zero operands with SVE compare
-			"cmeq p1.b, p0/z, z0.b, #0",          // a == 0
-			"cmeq p2.b, p0/z, z1.b, #0",          // b == 0
-			"orr p3.b, p0/z, p1.b, p2.b",         // either zero
-			
-			// Early exit if all lanes are zero
-			"ptest p0, p3.b",
-			"b.eq 2f",                              // Branch if no zeros
-			
-			// Perform SVE table lookups for log operations
-			// (This would need the lookup tables in SVE format)
-			"mov z2.b, #0",                        // Initialize result
-			"b 3f",                                 // Skip to end
-			
-			"2:",
-			// Regular multiplication path
-			"eor z2.d, z0.d, z1.d",               // XOR for GF(2) addition
-			
-			"3:",
-			// Extract result
-			"mov {result:x}, z2.d[0]",
-			
-			a_vec = in(reg) a_vec,
-			b_vec = in(reg) b_vec,
-			result = out(reg) result,
-			options(pure, nomem, nostack)
-		);
+		// Check if all lanes are zero for early return
+		if vmaxvq_u8(either_zero) == 0xFF {
+			return M128::from(vdupq_n_u8(0));
+		}
 		
 		// For complex operations, fall back to optimized NEON version
 		// Full SVE lookup table implementation would require restructuring tables
