+++ crates/ntt/src/single_threaded.rs
@@ -18,69 +18,68 @@ use crate::twiddle::{OnTheFlyTwiddleAccess, PrecomputedTwiddleAccess, expand_sub
 #[cfg(all(target_arch = "aarch64", target_feature = "sve"))]
 mod sve_ntt {
 	use super::*;
-	use std::arch::asm;
+	
 	
 	/// SVE-optimized forward butterfly operation
+	#[allow(dead_code)]
 	#[inline]
 	pub fn sve_forward_butterfly<P: PackedField<Scalar: BinaryField>>(
 		data: &mut [P],
 		stride: usize,
 		twiddle: P::Scalar,
 	) {
-		unsafe {
-			// SVE implementation for parallel butterfly operations
-			// This processes multiple butterfly units simultaneously
-			
-			let packed_twiddle = P::broadcast(twiddle);
-			
-			// Process data in SVE-sized chunks for maximum vectorization
-			for chunk in data.chunks_exact_mut(stride * 2) {
-				if chunk.len() >= stride * 2 {
-					let (left, right) = chunk.split_at_mut(stride);
-					
-					// SVE vectorized butterfly: (a, b) -> (a + b*t, b)
-					// where t is the twiddle factor
-					for (a, b) in left.iter_mut().zip(right.iter()) {
-						let scaled_b = *b * packed_twiddle;
-						*a += scaled_b;
-					}
+		// SVE implementation for parallel butterfly operations
+		// This processes multiple butterfly units simultaneously
+		
+		let packed_twiddle = P::broadcast(twiddle);
+		
+		// Process data in SVE-sized chunks for maximum vectorization
+		for chunk in data.chunks_exact_mut(stride * 2) {
+			if chunk.len() >= stride * 2 {
+				let (left, right) = chunk.split_at_mut(stride);
+				
+				// SVE vectorized butterfly: (a, b) -> (a + b*t, b)
+				// where t is the twiddle factor
+				for (a, b) in left.iter_mut().zip(right.iter()) {
+					let scaled_b = *b * packed_twiddle;
+					*a += scaled_b;
 				}
 			}
 		}
 	}
 	
 	/// SVE-optimized inverse butterfly operation
+	#[allow(dead_code)]
 	#[inline]
 	pub fn sve_inverse_butterfly<P: PackedField<Scalar: BinaryField>>(
 		data: &mut [P],
 		stride: usize,
 		twiddle: P::Scalar,
 	) {
-		unsafe {
-			// SVE implementation for parallel inverse butterfly operations
-			
-			let packed_twiddle = P::broadcast(twiddle);
-			
-			// Process data in SVE-sized chunks for maximum vectorization
-			for chunk in data.chunks_exact_mut(stride * 2) {
-				if chunk.len() >= stride * 2 {
-					let (left, right) = chunk.split_at_mut(stride);
-					
-					// SVE vectorized inverse butterfly: (a, b) -> (a - b*t, b)
-					for (a, b) in left.iter_mut().zip(right.iter()) {
-						let scaled_b = *b * packed_twiddle;
-						*a -= scaled_b;
-					}
+		// SVE implementation for parallel inverse butterfly operations
+		
+		let packed_twiddle = P::broadcast(twiddle);
+		
+		// Process data in SVE-sized chunks for maximum vectorization
+		for chunk in data.chunks_exact_mut(stride * 2) {
+			if chunk.len() >= stride * 2 {
+				let (left, right) = chunk.split_at_mut(stride);
+				
+				// SVE vectorized inverse butterfly: (a, b) -> (a - b*t, b)
+				for (a, b) in left.iter_mut().zip(right.iter()) {
+					let scaled_b = *b * packed_twiddle;
+					*a -= scaled_b;
 				}
 			}
 		}
 	}
 	
 	/// SVE-optimized batch NTT layer processing
+	#[allow(dead_code)]
 	#[inline]
 	pub fn sve_ntt_layer<F: BinaryField, P: PackedField<Scalar = F>>(
 		data: &mut [P],
-		shape: NTTShape,
+		_shape: NTTShape,
 		layer: usize,
 		s_evals: &impl TwiddleAccess<F>,
 		forward: bool,
